git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@123663 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: 2a0fa709c0bb1952d8853044f42f6204872f3a4d

+++ Online/Online/Gaucho/src/components/MEPSvc.cpp
@@ -159,7 +159,7 @@ void MEPSvc::analyze(void *, int ,MonMap* mmap)
     DetData_T<double>::iterator jj = ii->second.find("rxOct");
     if (jj!=ii->second.end())
     {
-      m_DataRate = jj->second;
+      m_DataRate = (float)jj->second;
       if (m_LHCbDataRate == 0)
       {
         m_LHCbDataRate = new DimService((char*)(m_PartitionName+"_DataRate").c_str(),m_DataRate);
